Move teliod configs to separate file
